added platform controller and helper to translate request messages

+++ Samples/PublicSamples/EchoBot/src/EchoBot.Api/Controllers/SynthesizerController.cs
@@ -60,6 +60,6 @@ namespace EchoBot.Api.Controllers
 
         public string? Style { get; set; }
 
-        public int? StyleDegree { get; set; }
+        public float? StyleDegree { get; set; }
     }
 }
